// src/api-client/models/revenue-by-month-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spring-Adam-Store
 * Api Service
 *
 * The version of the OpenAPI document: v1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface RevenueByMonthDTO
 */
export interface RevenueByMonthDTO {
    /**
     * 
     * @type {object}
     * @memberof RevenueByMonthDTO
     */
    'month'?: object;
    /**
     * 
     * @type {number}
     * @memberof RevenueByMonthDTO
     */
    'totalAmount'?: number;
}